requirements.txt
python-telegram-bot==21.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
